backend/requirements.txt
# Create project structure first:
# mkdir -p smart-education-platform/backend/api/{core,modules/parents,modules/auth}
# cd smart-education-platform/backend

# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]
passlib[bcrypt]
python-multipart